requirements.txt
asn1crypto==1.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
Django==2.2.5
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-uuidfield==0.5.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
openapi-codec==1.3.2
pycparser==2.19
pyfcm==1.4.5
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
requests-toolbelt==0.9.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
uWSGI==2.0.18
